refactor: login test
